+++ tests/profiles/test_igd.py
@@ -2,10 +2,10 @@
 
 import pytest
 
-from async_upnp_client import UpnpEventHandler, UpnpFactory
+from async_upnp_client import UpnpFactory
 from async_upnp_client.profiles.igd import IgdDevice
 
-from ..upnp_test_requester import RESPONSE_MAP, UpnpTestRequester, read_file
+from ..conftest import RESPONSE_MAP, UpnpTestNotifyServer, UpnpTestRequester, read_file
 
 
 @pytest.mark.asyncio
@@ -14,7 +14,11 @@ async def test_init_igd_profile() -> None:
     requester = UpnpTestRequester(RESPONSE_MAP)
     factory = UpnpFactory(requester)
     device = await factory.async_create_device("http://igd:1234/device.xml")
-    event_handler = UpnpEventHandler("http://localhost:11302", requester)
+    notify_server = UpnpTestNotifyServer(
+        requester=requester,
+        source=("192.168.1.2", 8090),
+    )
+    event_handler = notify_server.event_handler
     profile = IgdDevice(device, event_handler=event_handler)
     assert profile
 
@@ -31,7 +35,11 @@ async def test_get_total_bytes_received() -> None:
     requester = UpnpTestRequester(responses)
     factory = UpnpFactory(requester)
     device = await factory.async_create_device("http://igd:1234/device.xml")
-    event_handler = UpnpEventHandler("http://localhost:11302", requester)
+    notify_server = UpnpTestNotifyServer(
+        requester=requester,
+        source=("192.168.1.2", 8090),
+    )
+    event_handler = notify_server.event_handler
     profile = IgdDevice(device, event_handler=event_handler)
     total_bytes_received = await profile.async_get_total_bytes_received()
     assert total_bytes_received == 1337
@@ -49,7 +57,11 @@ async def test_get_total_packets_received_empty_response() -> None:
     requester = UpnpTestRequester(responses)
     factory = UpnpFactory(requester)
     device = await factory.async_create_device("http://igd:1234/device.xml")
-    event_handler = UpnpEventHandler("http://localhost:11302", requester)
+    notify_server = UpnpTestNotifyServer(
+        requester=requester,
+        source=("192.168.1.2", 8090),
+    )
+    event_handler = notify_server.event_handler
     profile = IgdDevice(device, event_handler=event_handler)
     total_bytes_received = await profile.async_get_total_packets_received()
     assert total_bytes_received is None
@@ -67,7 +79,11 @@ async def test_get_status_info_invalid_uptime() -> None:
     requester = UpnpTestRequester(responses)
     factory = UpnpFactory(requester)
     device = await factory.async_create_device("http://igd:1234/device.xml")
-    event_handler = UpnpEventHandler("http://localhost:11302", requester)
+    notify_server = UpnpTestNotifyServer(
+        requester=requester,
+        source=("192.168.1.2", 8090),
+    )
+    event_handler = notify_server.event_handler
     profile = IgdDevice(device, event_handler=event_handler)
     status_info = await profile.async_get_status_info()
     assert status_info is None